Module bc_hash::sha256

Structs

Digest

Represents a SHA-256 digest in binary format.

Constants

CONSTANTS

An array of 64 constants consisting of the first 32 bits of the fractional parts of the cube roots of the first 64 primes 2 through 311.

DIGEST_SIZE

The total size of a digest object’s data array in bytes.

INITIAL_VALUES

An array used to initialize a digest to the first 32 bits of the fractional parts of the square roots of the first 8 primes, 2 through 19.